+++ utils/transforms.js
@@ -1,0 +1,12 @@
+const htmlmin = require('html-minifier')
+
+module.exports = {
+  htmlmin: function (content, outputPath) {
+    return htmlmin.minify(content, {
+      useShortDoctype: true,
+      removeComments: true,
+      collapseWhitespace: true
+    })
+    return content
+  },
+}